r/virtual_machine Add `vapp_transport`
